re-Push of Soundbank test (Opening)
Note: There are duplicate assets and play calls as this is really a test for me and Reef to test a theory.

The theory is that a Unity game call cannot pull from multiple soundbanks at a time and we are testing creating an Opening Soundbank to separate soundbanks by timeline usage rather than VO vs Music.