Play_THEMATIC_SAVASANA_SEQUENCE
Eliminated the need for a Silent Music meditation by adding internally triggered Wwise behavior to initiate Play_WakeUpEndSoon_Sequence after 2 minutes. This is the only game call necessary for the CLOSING bank and should be initiated according to programatic timing. It will go all the way through to the Exit Music at 45:00. Tested and approved in Wwise.